+++ components/admin/JsonUpload/JsonUpload.stories.tsx
@@ -10,29 +10,27 @@ export default {
 export const defaultJSONUpload = (): React.ReactElement => <JSONUpload></JSONUpload>;
 
 const testForm = {
-  formID: "test0form00000id000asdf11",
-  formConfig: {
-    publishingStatus: true,
-    securityAttribute: "Unclassified",
-    submission: {
-      email: "test@test.com",
-    },
-    form: {
-      version: "1",
-      titleEn: "Test JSON!",
-      titleFr: "Test JSON!",
-      layout: [1],
-      elements: [
-        {
-          id: 1,
-          type: FormElementTypes.textField,
-          properties: {
-            titleEn: "test Element!",
-            titleFr: "test Element!",
-          },
+  id: "test0form00000id000asdf11",
+  publishingStatus: true,
+  securityAttribute: "Unclassified",
+  submission: {
+    email: "test@test.com",
+  },
+  form: {
+    version: "1",
+    titleEn: "Test JSON!",
+    titleFr: "Test JSON!",
+    layout: [1],
+    elements: [
+      {
+        id: 1,
+        type: FormElementTypes.textField,
+        properties: {
+          titleEn: "test Element!",
+          titleFr: "test Element!",
         },
-      ],
-    },
+      },
+    ],
   },
 };