+++ components/form-builder/elements/RichTextLocked.tsx
@@ -1,10 +1,12 @@
 import React, { useRef, useEffect, useState } from "react";
 import styled from "styled-components";
 import useTemplateStore from "../store/useTemplateStore";
-import { RichTextEditor } from "../editor/RichTextEditor";
-import { Descendant } from "slate";
-import { serialize } from "../editor/Markdown";
+import { RichTextEditor } from "../plate-editor/RichTextEditor";
+import { serializeMd } from "../plate-editor/helpers/markdown";
+import { deserializeMd, Value } from "@udecode/plate";
 import { PanelActionsLocked } from "../panel/PanelActionsLocked";
+import { LocalizedElementProperties } from "../types";
+import { useMyPlateEditorRef } from "../plate-editor/types";
 
 const ElementWrapperDiv = styled.div`
   border: 1.5px solid #000000;
@@ -17,37 +19,45 @@ const ElementWrapperDiv = styled.div`
 
 const ContentWrapper = styled.div`
   display: flex;
-  margin: 0px 20px 10px;
+  margin: 0px 20px;
 
-  & h3 {
-    margin-top: 30px;
+  & h2 {
+    font-size: 26px;
+    line-height: 32px;
+    margin-top: 15px;
     margin-bottom: 0;
+    padding-bottom: 0;
   }
 `;
 
 const OptionWrapper = styled.div`
   display: flex;
-  margin: 30px 20px 10px;
-  & div {
-    width: 100%;
-    margin-bottom: 20px;
-  }
+  margin: 0 26px;
 `;
 
 export const RichTextLocked = ({
+  id,
   addElement,
   children,
   initialValue,
   schemaProperty,
+  "aria-label": ariaLabel = undefined,
 }: {
+  id: string;
   addElement: boolean;
   children?: React.ReactElement;
-  initialValue: Descendant[];
+  initialValue: string;
   schemaProperty: string;
+  "aria-label"?: string;
 }) => {
   const input = useRef<HTMLInputElement>(null);
-  const { updateField } = useTemplateStore();
-  const [value, setValue] = useState<Descendant[]>(initialValue);
+  const { localizeField, updateField } = useTemplateStore();
+  const editorId = `${id}-editor`;
+  const editor = useMyPlateEditorRef();
+
+  const [value, setValue] = useState(
+    initialValue ? deserializeMd(editor, initialValue) : [{ children: [{ text: "" }] }]
+  );
 
   useEffect(() => {
     if (input.current) {
@@ -55,18 +65,26 @@ export const RichTextLocked = ({
     }
   }, []);
 
-  const onChange = (value: string) => {
-    const parsed = JSON.parse(value);
-    const serialized = serialize({ children: parsed });
-    setValue(parsed);
-    updateField(`form.${schemaProperty}.descriptionEn`, serialized);
+  const onChange = (value: Value) => {
+    let serialized = serializeMd(value);
+
+    if (typeof serialized === "undefined") {
+      serialized = "";
+    }
+
+    setValue(value);
+
+    updateField(
+      `form.${schemaProperty}.${localizeField(LocalizedElementProperties.DESCRIPTION)}`,
+      serialized
+    );
   };
 
   return (
     <ElementWrapperDiv>
       <ContentWrapper>{children}</ContentWrapper>
       <OptionWrapper>
-        <RichTextEditor value={value} onChange={onChange} />
+        <RichTextEditor id={editorId} value={value} onChange={onChange} aria-label={ariaLabel} />
       </OptionWrapper>
       <PanelActionsLocked addElement={addElement} />
     </ElementWrapperDiv>