+++ components/form-builder/icons/ShortAnswerIcon.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 export const ShortAnswerIcon = () => (
-  <svg width={24} height={24} fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M4 5h16v2H4V5Zm0 4h10v2H4V9Z" fill="#5F6368" />
+  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M4 15v-2h10v2Zm0-4V9h16v2Z" fill="#5F6368" />
   </svg>
 );