+++ components/form-builder/layout/Preview.tsx
@@ -10,8 +10,8 @@ export const Preview = () => {
   const stringified = getSchema();
 
   const formRecord = {
-    formID: "test0form00000id000asdf11",
-    formConfig: JSON.parse(stringified),
+    id: "test0form00000id000asdf11",
+    ...JSON.parse(stringified),
   };
   const router = useRouter();
   const { t, i18n } = useTranslation();
@@ -20,10 +20,8 @@ export const Preview = () => {
   const currentForm = getRenderedForm(formRecord, language, t);
 
   return (
-    <>
-      <Form formRecord={formRecord} language={language} router={router} t={t} isPreview={true}>
-        {currentForm}
-      </Form>
-    </>
+    <Form formRecord={formRecord} language={language} router={router} t={t} isPreview={true}>
+      {currentForm}
+    </Form>
   );
 };