+++ components/form-builder/panel/ElementPanel.tsx
@@ -6,42 +6,20 @@ import useTemplateStore from "../store/useTemplateStore";
 import useModalStore from "../store/useModalStore";
 import { Select } from "../elements";
 import { PanelActions } from "./PanelActions";
-import { ElementOption, ElementProperties, ElementTypeWithIndex } from "../types";
+import {
+  ElementOption,
+  ElementProperties,
+  ElementTypeWithIndex,
+  LocalizedElementProperties,
+} from "../types";
 import { UseSelectStateChange } from "downshift";
 import { ShortAnswer, Options, RichText, RichTextLocked } from "../elements";
-import {
-  CalendarIcon,
-  CheckBoxEmptyIcon,
-  CheckIcon,
-  EmailIcon,
-  ParagraphIcon,
-  PhoneIcon,
-  RadioIcon,
-  RadioEmptyIcon,
-  SelectMenuIcon,
-  ShortAnswerIcon,
-} from "../icons";
+import { useElementOptions } from "../hooks/useElementOptions";
+import { CheckBoxEmptyIcon, RadioEmptyIcon } from "../icons";
 import { ModalButton } from "./Modal";
 import { Checkbox } from "./MultipleChoice";
 import { FancyButton } from "./Button";
 import { Input } from "./Input";
-
-const Separator = styled.div`
-  border-top: 1px solid rgba(0, 0, 0, 0.12);
-  margin: 8px 0;
-`;
-
-const elementOptions = [
-  { id: "textField", value: "Short answer", icon: <ShortAnswerIcon /> },
-  { id: "richText", value: "Rich Text", icon: <ParagraphIcon /> },
-  { id: "textArea", value: "Paragraph", icon: <ParagraphIcon />, prepend: <Separator /> },
-  { id: "radio", value: "Multiple choice", icon: <RadioIcon /> },
-  { id: "checkbox", value: "Checkboxes", icon: <CheckIcon /> },
-  { id: "dropdown", value: "Dropdown", icon: <SelectMenuIcon />, prepend: <Separator /> },
-  { id: "email", value: "Email", icon: <EmailIcon /> },
-  { id: "phone", value: "Phone number", icon: <PhoneIcon /> },
-  { id: "date", value: "Date", icon: <CalendarIcon /> },
-];
 
 const SelectedElement = ({
   selected,
@@ -55,6 +33,7 @@ const SelectedElement = ({
   let element = null;
 
   switch (selected.id) {
+    case "text":
     case "textField":
       element = <ShortAnswer>{t("Short answer text")}</ShortAnswer>;
       break;
@@ -82,6 +61,9 @@ const SelectedElement = ({
     case "date":
       element = <ShortAnswer>mm/dd/yyyy</ShortAnswer>;
       break;
+    case "number":
+      element = <ShortAnswer>0123456789</ShortAnswer>;
+      break;
     default:
       element = null;
   }
@@ -90,6 +72,7 @@ const SelectedElement = ({
 };
 
 const getSelectedOption = (item: ElementTypeWithIndex): ElementOption => {
+  const elementOptions = useElementOptions();
   const {
     form: { elements },
   } = useTemplateStore();
@@ -98,13 +81,15 @@ const getSelectedOption = (item: ElementTypeWithIndex): ElementOption => {
   if (!type) {
     return elementOptions[2];
   } else if (type === "textField") {
+    const validationType = elements[item.index].properties.validation.type;
     /**
      * Email, phone, and date fields are specialized text field types.
      * That is to say, their "type" is "textField" but they have specalized validation "type"s.
      * So if we have a "textField", we want to first check properties.validation.type to see if
      * it is a true Short Answer, or one of the other types.
+     * The one exception to this is validationType === "text" types, for which we want to return "textField"
      */
-    type = elements[item.index].properties.validation.type || type;
+    type = validationType && validationType !== "text" ? validationType : type;
   }
 
   const selected = elementOptions.filter((item) => item.id === type);
@@ -119,12 +104,12 @@ const Row = styled.div<RowProps>`
   display: flex;
   justify-content: space-between;
   position: relative;
-
-  & div {
+  & > div {
     ${({ isRichText }) =>
       isRichText &&
       `
       width: 100%;
+      margin: 0;
     `}
   }
 `;
@@ -216,18 +201,23 @@ const QuestionNumber = styled.span`
 const Form = ({ item }: { item: ElementTypeWithIndex }) => {
   const isRichText = item.type == "richText";
   const { t } = useTranslation("form-builder");
+  const elementOptions = useElementOptions();
   const {
+    localizeField,
     form: { elements },
     updateField,
     unsetField,
     resetChoices,
+    focusInput,
+    setFocusInput,
   } = useTemplateStore();
 
   const input = useRef<HTMLInputElement>(null);
 
   useEffect(() => {
-    if (input.current) {
+    if (input.current && focusInput) {
       input.current.focus();
+      setFocusInput(false);
     }
   }, []);
 
@@ -240,11 +230,20 @@ const Form = ({ item }: { item: ElementTypeWithIndex }) => {
 
   const _updateState = (id: string, index: number) => {
     switch (id) {
+      case "text":
+      case "textField":
       case "email":
       case "phone":
       case "date":
+      case "number":
         updateField(`form.elements[${index}].type`, "textField");
-        updateField(`form.elements[${index}].properties.validation.type`, id);
+
+        if (id === "textField" || id === "text") {
+          unsetField(`form.elements[${index}].properties.validation.type`);
+        } else {
+          updateField(`form.elements[${index}].properties.validation.type`, id);
+          unsetField(`form.elements[${index}].properties.validation.maxLength`);
+        }
         break;
       case "richText":
         resetChoices(index);
@@ -252,6 +251,7 @@ const Form = ({ item }: { item: ElementTypeWithIndex }) => {
       default: // eslint-disable-line no-fallthrough
         updateField(`form.elements[${index}].type`, id);
         unsetField(`form.elements[${index}].properties.validation.type`);
+        unsetField(`form.elements[${index}].properties.validation.maxLength`);
         break;
     }
   };
@@ -266,6 +266,8 @@ const Form = ({ item }: { item: ElementTypeWithIndex }) => {
     [setSelectedItem]
   );
 
+  const hasDescription = item.properties[localizeField(LocalizedElementProperties.DESCRIPTION)];
+
   return (
     <>
       <Row isRichText={isRichText}>
@@ -277,19 +279,34 @@ const Form = ({ item }: { item: ElementTypeWithIndex }) => {
               <TitleInput
                 ref={input}
                 type="text"
+                id={`item${item.index}`}
                 name={`item${item.index}`}
                 placeholder={t("Question")}
-                value={item.properties.titleEn}
+                value={item.properties[localizeField(LocalizedElementProperties.TITLE)]}
+                aria-describedby={hasDescription ? `item${item.index}-describedby` : undefined}
                 onChange={(e) => {
-                  updateField(`form.elements[${item.index}].properties.titleEn`, e.target.value);
+                  updateField(
+                    `form.elements[${item.index}].properties.${localizeField(
+                      LocalizedElementProperties.TITLE
+                    )}`,
+                    e.target.value
+                  );
                 }}
               />
             </>
           )}
-          {item.properties.descriptionEn && item.type !== "richText" && (
-            <DivDisabled aria-label={t("Description")}>{item.properties.descriptionEn}</DivDisabled>
+          {hasDescription && item.type !== "richText" && (
+            <DivDisabled id={`item${item.index}-describedby`}>
+              {item.properties[localizeField(LocalizedElementProperties.DESCRIPTION)]}
+            </DivDisabled>
           )}
           <SelectedElement item={item} selected={selectedItem} />
+          {item.properties.validation.maxLength && (
+            <DivDisabled>
+              {t("Max character length: ")}
+              {item.properties.validation.maxLength}
+            </DivDisabled>
+          )}
         </div>
         {!isRichText && (
           <>
@@ -344,6 +361,10 @@ const ModalInput = styled(Input)`
   width: 90%;
 `;
 
+const ModalInputShort = styled(Input)`
+  width: 180px;
+`;
+
 const ModalSaveButton = styled(FancyButton)`
   padding: 15px 20px;
   background: #26374a;
@@ -367,12 +388,15 @@ const ModalForm = ({
   item,
   properties,
   updateModalProperties,
+  unsetModalField,
 }: {
   item: ElementTypeWithIndex;
   properties: ElementProperties;
   updateModalProperties: (index: number, properties: ElementProperties) => void;
+  unsetModalField: (path: string) => void;
 }) => {
   const { t } = useTranslation("form-builder");
+  const { localizeField } = useTemplateStore();
 
   return (
     <form onSubmit={(e: React.FormEvent<HTMLFormElement>) => e.preventDefault()}>
@@ -383,11 +407,11 @@ const ModalForm = ({
           type="text"
           name={`item${item.index}`}
           placeholder={t("Question")}
-          value={properties.titleEn}
+          value={properties[localizeField(LocalizedElementProperties.TITLE)]}
           onChange={(e) =>
             updateModalProperties(item.index, {
               ...properties,
-              ...{ titleEn: e.target.value },
+              ...{ local: e.target.value },
             })
           }
         />
@@ -404,10 +428,10 @@ const ModalForm = ({
           onChange={(e) => {
             updateModalProperties(item.index, {
               ...properties,
-              ...{ descriptionEn: e.target.value },
+              ...{ [localizeField(LocalizedElementProperties.DESCRIPTION)]: e.target.value },
             });
           }}
-          value={properties.descriptionEn}
+          value={properties[localizeField(LocalizedElementProperties.DESCRIPTION)]}
         />
       </ModalRow>
       <ModalRow>
@@ -419,14 +443,61 @@ const ModalForm = ({
           value={`required-${item.index}-value-modal`}
           checked={properties.validation.required}
           onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+            // clone the existing properties so that we don't overwrite other keys in "validation"
+            const validation = Object.assign({}, properties.validation, {
+              required: e.target.checked,
+            });
             updateModalProperties(item.index, {
               ...properties,
-              ...{ validation: { required: e.target.checked } },
+              ...{ validation },
             });
           }}
           label={t("Required")}
         ></Checkbox>
       </ModalRow>
+      {item.type === "textField" &&
+        (!item.properties.validation.type || item.properties.validation.type === "text") && (
+          <ModalRow>
+            <FormLabel htmlFor={`characterLength--modal--${item.index}`}>
+              {t("Maximum character length")}
+            </FormLabel>
+            <HintText>
+              {t(
+                "Only use a character limit when there is a good reason for limiting the number of characters users can enter."
+              )}
+            </HintText>
+            <ModalInputShort
+              id={`characterLength--modal--${item.index}`}
+              type="number"
+              min="1"
+              value={properties.validation.maxLength || ""}
+              onKeyDown={(e) => {
+                if (["-", "+", ".", "e"].includes(e.key)) {
+                  e.preventDefault();
+                }
+              }}
+              onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+                // if value is "", unset the field
+                if (e.target.value === "") {
+                  unsetModalField(`modals[${item.index}].validation.maxLength`);
+                  return;
+                }
+
+                const value = parseInt(e.target.value);
+                if (!isNaN(value) && value >= 1) {
+                  // clone the existing properties so that we don't overwrite other keys in "validation"
+                  const validation = Object.assign({}, properties.validation, {
+                    maxLength: value,
+                  });
+                  updateModalProperties(item.index, {
+                    ...properties,
+                    ...{ validation },
+                  });
+                }
+              }}
+            />
+          </ModalRow>
+        )}
     </form>
   );
 };
@@ -453,7 +524,7 @@ export const ElementWrapper = ({ item }: { item: ElementTypeWithIndex }) => {
     updateField,
   } = useTemplateStore();
 
-  const { isOpen, modals, updateModalProperties } = useModalStore();
+  const { isOpen, modals, updateModalProperties, unsetModalField } = useModalStore();
 
   React.useEffect(() => {
     if (item.type != "richText") {
@@ -492,6 +563,7 @@ export const ElementWrapper = ({ item }: { item: ElementTypeWithIndex }) => {
             item={item}
             properties={modals[item.index]}
             updateModalProperties={updateModalProperties}
+            unsetModalField={unsetModalField}
           />
         )}
       </PanelActions>
@@ -513,57 +585,50 @@ const ElementPanelDiv = styled.div`
 `;
 
 export const ElementPanel = () => {
-  const {
-    form: { elements },
-  } = useTemplateStore();
+  const { t } = useTranslation("form-builder");
+  const { form, localizeField } = useTemplateStore();
 
-  const introTextPlaceholder = [
-    {
-      type: "paragraph",
-      children: [{ text: "Add an introduction" }],
-    },
-  ];
+  const introTextPlaceholder =
+    form.introduction[localizeField(LocalizedElementProperties.DESCRIPTION)];
 
-  const confirmTextPlaceholder = [
-    {
-      type: "paragraph",
-      children: [{ text: "Thank you for participating in the ice cream survey!" }],
-    },
-  ];
+  const confirmTextPlaceholder =
+    form.endPage[localizeField(LocalizedElementProperties.DESCRIPTION)];
 
-  const policyTextPlaceholder = [
-    {
-      type: "paragraph",
-      children: [{ text: "Email addresses will not be shared." }],
-    },
-  ];
+  const policyTextPlaceholder =
+    form.privacyPolicy[localizeField(LocalizedElementProperties.DESCRIPTION)];
 
   return (
     <ElementPanelDiv>
       <RichTextLocked
+        id="introductionPage"
         addElement={true}
         initialValue={introTextPlaceholder}
         schemaProperty="introduction"
+        aria-label={t("richTextIntroTitle")}
       />
-      {elements.map((element, index) => {
+      {form.elements.map((element, index) => {
         const item = { ...element, index };
         return <ElementWrapper item={item} key={item.id} />;
       })}
-      {elements?.length >= 1 && (
+      {form.elements?.length >= 1 && (
         <>
           <RichTextLocked
+            id="endPage"
             addElement={false}
             initialValue={confirmTextPlaceholder}
             schemaProperty="endPage"
+            aria-label={t("richTextConfirmationTitle")}
           >
-            <h3>Confirmation page and message</h3>
+            <h2>{t("richTextConfirmationTitle")}</h2>
           </RichTextLocked>
           <RichTextLocked
+            id="policyPage"
             addElement={false}
             initialValue={policyTextPlaceholder}
             schemaProperty="privacyPolicy"
+            aria-label={t("richTextPrivacyTitle")}
           >
-            <h3>Privacy statement</h3>
+            <h2>{t("richTextPrivacyTitle")}</h2>
           </RichTextLocked>
         </>
       )}