+++ components/globals/Footer.js
@@ -10,14 +10,7 @@ const Footer = ({ disableGcBranding }) => {
     <footer className="gc-footer" data-testid="footer">
       <div className="gc-footer-container items-start">
         <div>
-          {!isSplashPage() && (
-            <>
-              {t("footer.terms.disclaimer")}
-              <br />
-              {t("footer.terms.desc-preface")}
-              <a href={t("footer.terms.link")}>{t("footer.terms.desc")}</a>.
-            </>
-          )}
+          {!isSplashPage() && <a href={t("footer.terms.link")}>{t("footer.terms.desc")}</a>}
         </div>
         {!disableGcBranding && (
           <div>