+++ lib/tests/templates.test.ts
@@ -10,7 +10,8 @@ import {
   getSubmissionTypeByID,
   onlyIncludePublicProperties,
 } from "../templates";
-import { FormConfiguration, FormRecord } from "@lib/types/form-types";
+
+import { BetterOmit, FormRecord } from "@lib/types";
 import formConfiguration from "@jestFixtures/cdsIntakeTestForm.json";
 
 // structuredClone is available starting in Node 17.
@@ -41,18 +42,19 @@ describe("Template CRUD functions", () => {
       jsonConfig: formConfiguration,
     });
 
-    const newTemplate = await createTemplate(formConfiguration as FormConfiguration);
-
-    expect(
-      prismaMock.template.create.calledWith({
-        data: {
-          jsonConfig: formConfiguration,
-        },
-      })
+    const newTemplate = await createTemplate(
+      formConfiguration as BetterOmit<FormRecord, "id" | "bearerToken">
     );
+
+    expect(prismaMock.template.create).toHaveBeenCalledWith({
+      data: {
+        jsonConfig: formConfiguration,
+      },
+    });
+
     expect(newTemplate).toEqual({
-      formID: "formtestID",
-      formConfig: formConfiguration,
+      id: "formtestID",
+      ...formConfiguration,
     });
   });
   test("Get a single Template", async () => {
@@ -62,21 +64,20 @@ describe("Template CRUD functions", () => {
     });
 
     const template = await getTemplateByID("formTestID");
-    expect(
-      prismaMock.template.findUnique.calledWith({
-        where: {
-          id: "formTestID",
-        },
-        select: {
-          id: true,
-          jsonConfig: true,
-        },
-      })
-    );
+
+    expect(prismaMock.template.findUnique).toHaveBeenCalledWith({
+      where: {
+        id: "formTestID",
+      },
+      select: {
+        id: true,
+        jsonConfig: true,
+      },
+    });
 
     expect(template).toEqual({
-      formID: "formtestID",
-      formConfig: formConfiguration,
+      id: "formtestID",
+      ...formConfiguration,
     });
   });
   test("Null returned when Template does not Exist", async () => {
@@ -100,12 +101,12 @@ describe("Template CRUD functions", () => {
     const templates = await getAllTemplates();
     expect(templates).toEqual([
       {
-        formID: "formtestID",
-        formConfig: formConfiguration,
+        id: "formtestID",
+        ...formConfiguration,
       },
       {
-        formID: "formtestID2",
-        formConfig: formConfiguration,
+        id: "formtestID2",
+        ...formConfiguration,
       },
     ]);
   });
@@ -116,7 +117,9 @@ describe("Template CRUD functions", () => {
     expect(template).toEqual([]);
   });
   test("Update Template", async () => {
-    const updatedFormConfig = structuredClone(formConfiguration as FormConfiguration);
+    const updatedFormConfig = structuredClone(
+      formConfiguration as BetterOmit<FormRecord, "id" | "bearerToken">
+    );
     updatedFormConfig.publishingStatus = true;
     (prismaMock.template.update as jest.MockedFunction<any>).mockResolvedValue({
       id: "formtestID",
@@ -125,24 +128,22 @@ describe("Template CRUD functions", () => {
 
     const updatedTemplate = await updateTemplate("test1", updatedFormConfig);
 
-    expect(
-      prismaMock.template.update.calledWith({
-        where: {
-          id: "formtestID",
-        },
-        data: {
-          jsonConfig: updatedFormConfig as unknown as Prisma.JsonObject,
-        },
-        select: {
-          id: true,
-          jsonConfig: true,
-        },
-      })
-    );
+    expect(prismaMock.template.update).toHaveBeenCalledWith({
+      where: {
+        id: "test1",
+      },
+      data: {
+        jsonConfig: updatedFormConfig as unknown as Prisma.JsonObject,
+      },
+      select: {
+        id: true,
+        jsonConfig: true,
+      },
+    });
 
     expect(updatedTemplate).toEqual({
-      formID: "formtestID",
-      formConfig: updatedFormConfig,
+      id: "formtestID",
+      ...updatedFormConfig,
     });
   });
   test("Delete template", async () => {
@@ -152,20 +153,26 @@ describe("Template CRUD functions", () => {
     });
 
     const deletedTemplate = await deleteTemplate("formtestID");
-    expect(
-      prismaMock.template.delete.calledWith({
-        where: {
-          id: "formtestID",
-        },
-      })
-    );
+
+    expect(prismaMock.template.delete).toHaveBeenCalledWith({
+      where: {
+        id: "formtestID",
+      },
+      select: {
+        id: true,
+        jsonConfig: true,
+      },
+    });
+
     expect(deletedTemplate).toEqual({
-      formID: "formtestID",
-      formConfig: formConfiguration,
+      id: "formtestID",
+      ...formConfiguration,
     });
   });
   test("Get Templates by publishing status", async () => {
-    const updatedFormConfig = structuredClone(formConfiguration as FormConfiguration);
+    const updatedFormConfig = structuredClone(
+      formConfiguration as BetterOmit<FormRecord, "id" | "bearerToken">
+    );
     updatedFormConfig.publishingStatus = true;
     (prismaMock.template.findMany as jest.MockedFunction<any>).mockResolvedValue([
       {
@@ -182,13 +189,11 @@ describe("Template CRUD functions", () => {
     expect(publishedTemplates).toHaveLength(1);
     expect(publishedTemplates).toMatchObject([
       {
-        formID: "formtestID2",
-        formConfig: {
-          publishingStatus: true,
-          displayAlphaBanner: true,
-          securityAttribute: "Unclassified",
-          form: updatedFormConfig.form,
-        },
+        id: "formtestID2",
+        publishingStatus: true,
+        displayAlphaBanner: true,
+        securityAttribute: "Unclassified",
+        form: updatedFormConfig.form,
       },
     ]);
   });
@@ -198,30 +203,29 @@ describe("Template CRUD functions", () => {
     });
 
     const submissionType = await getSubmissionTypeByID("formtestID");
-    expect(
-      prismaMock.template.findUnique.calledWith({
-        where: {
-          id: "formtestID",
-        },
-        select: {
-          jsonConfig: true,
-        },
-      })
-    );
+
+    expect(prismaMock.template.findUnique).toHaveBeenCalledWith({
+      where: {
+        id: "formtestID",
+      },
+      select: {
+        jsonConfig: true,
+      },
+    });
 
     expect(submissionType).toEqual(formConfiguration.submission);
   });
   test("Only include public properties", async () => {
     const formRecord = {
-      formID: "testID",
-      formConfig: formConfiguration,
+      id: "testID",
+      ...formConfiguration,
     };
 
     const publicFormRecord = onlyIncludePublicProperties(formRecord as FormRecord);
-    expect(publicFormRecord.formConfig).not.toHaveProperty("submission");
-    expect(publicFormRecord.formConfig).not.toHaveProperty("internalTitleEn");
-    expect(publicFormRecord.formConfig).not.toHaveProperty("internalTitleFr");
-    expect(publicFormRecord.formConfig).toHaveProperty("displayAlphaBanner");
-    expect(publicFormRecord.formConfig).toHaveProperty("securityAttribute");
+    expect(publicFormRecord).not.toHaveProperty("submission");
+    expect(publicFormRecord).not.toHaveProperty("internalTitleEn");
+    expect(publicFormRecord).not.toHaveProperty("internalTitleFr");
+    expect(publicFormRecord).toHaveProperty("displayAlphaBanner");
+    expect(publicFormRecord).toHaveProperty("securityAttribute");
   });
 });